+++ packages/bpk-component-accordion/index.js
@@ -1,0 +1,5 @@
+import BpkAccordion from './src/BpkAccordion';
+import BpkAccordionContainer from './src/BpkAccordionContainer';
+import BpkAccordionItem from './src/BpkAccordionItem';
+
+export { BpkAccordionContainer, BpkAccordion, BpkAccordionItem };