+++ packages/bpk-component-paragraph/src/BpkParagraph.js
@@ -2,13 +2,22 @@ import React, { PropTypes } from 'react';
 
 import './bpk-paragraph.scss';
 
-const BpkParagraph = props => <p className="bpk-paragraph">{props.children}</p>;
+const BpkParagraph = (props) => {
+  const classNames = ['bpk-paragraph'];
+  const { className, ...rest } = props;
+
+  if (className) { classNames.push(className); }
+
+  return <p className={classNames.join(' ')} {...rest} />;
+};
 
 BpkParagraph.propTypes = {
-  children: PropTypes.oneOfType([
-    PropTypes.arrayOf(PropTypes.node),
-    PropTypes.node,
-  ]).isRequired,
+  children: PropTypes.node.isRequired,
+  className: PropTypes.string,
+};
+
+BpkParagraph.defaultProps = {
+  className: null,
 };
 
 export default BpkParagraph;