+++ packages/bpk-component-radio/src/BpkRadio.js
@@ -4,8 +4,10 @@ import './bpk-radio.scss';
 
 const BpkRadio = (props) => {
   const classNames = ['bpk-radio'];
+  const { name, onChange, value, checked, label, disabled, className, ...rest } = props;
 
-  if (props.disabled) { classNames.push('bpk-radio--disabled'); }
+  if (disabled) { classNames.push('bpk-radio--disabled'); }
+  if (className) { classNames.push(className); }
 
   // This is awkward because the label-has-for rule enforces an 'id' / 'for' pairing
   // when it's not really necessary for nested inputs.
@@ -16,14 +18,14 @@ const BpkRadio = (props) => {
       <input
         type="radio"
         className="bpk-radio__input"
-        name={props.name}
-        onChange={props.onChange}
-        id={props.id}
-        value={props.value || props.name}
-        checked={props.checked}
-        disabled={props.disabled}
+        name={name}
+        onChange={onChange}
+        value={value || name}
+        checked={checked}
+        disabled={disabled}
+        {...rest}
       />
-      {props.label ? ` ${props.label}` : null}
+      {label ? ` ${label}` : null}
     </label>
   );
   /* eslint-enable */
@@ -32,19 +34,19 @@ const BpkRadio = (props) => {
 BpkRadio.propTypes = {
   name: PropTypes.string.isRequired,
   onChange: PropTypes.func.isRequired,
-  id: PropTypes.string,
   label: PropTypes.string,
   value: PropTypes.string,
   checked: PropTypes.bool,
   disabled: PropTypes.bool,
+  className: PropTypes.string,
 };
 
 BpkRadio.defaultProps = {
-  id: null,
   label: null,
   value: null,
   checked: false,
   disabled: false,
+  className: null,
 };
 
 export default BpkRadio;