+++ packages/bpk-docs/src/pages/AccordionsPage/index.js
@@ -1,0 +1,3 @@
+import page from './AccordionsPage';
+
+export default page;